Add nprogress as loading bar.
This makes every view in a router initiate loading before being rendered. The loading must be manually stopped in the component when elements are finished loading. An instance method for stopping loading is accessed through `this.$loaded()`, which also returns true for convenience should a data variable be updated at the same time.

Loading operations should be placed in a separate method so that ending loading can be performed more conveniently in case of multiple exit points in the loading operation. This new method can also be used later if there is a need for refreshing a component without re-rendering (e.g. if the router switches a property, but not components).

This commit applies the aforementioned changes to every view, excluding the Navbar (this is not a view, and so loading never starts for it).

Some refactoring is also included.